+++ dotnet/src/Encryptor.cs
@@ -233,6 +233,102 @@ namespace Microsoft.Research.SEAL
         }
 
         /// <summary>
+        /// Encrypts a plaintext with the public key and returns the ciphertext as
+        /// a serializable object.
+        /// </summary>
+        /// <remarks>
+        /// <para>
+        /// Encrypts a plaintext with the public key and returns the ciphertext as
+        /// a serializable object.
+        /// </para>
+        /// <para>
+        /// The encryption parameters for the resulting ciphertext correspond to:
+        /// 1) in BFV, the highest (data) level in the modulus switching chain,
+        /// 2) in CKKS, the encryption parameters of the plaintext.
+        /// Dynamic memory allocations in the process are allocated from the memory
+        /// pool pointed to by the given MemoryPoolHandle.
+        /// </para>
+        /// </remarks>
+        /// <param name="plain">The plaintext to encrypt</param>
+        /// <param name="pool">The MemoryPoolHandle pointing to a valid memory pool</param>
+        /// <exception cref="ArgumentNullException">if plain is null</exception>
+        /// <exception cref="InvalidOperationException">if a public key is not set</exception>
+        /// <exception cref="ArgumentException">if plain is not valid for the encryption
+        /// parameters</exception>
+        /// <exception cref="ArgumentException">if plain is not in default NTT
+        /// form</exception>
+        /// <exception cref="ArgumentException">if pool is uninitialized</exception>
+        public Serializable<Ciphertext> Encrypt(
+            Plaintext plain,
+            MemoryPoolHandle pool = null)
+        {
+            Ciphertext destination = new Ciphertext();
+            Encrypt(plain, destination, pool);
+            return new Serializable<Ciphertext>(destination);
+        }
+
+        /// <summary>
+        /// Encrypts a zero plaintext with the public key and returns the ciphertext
+        /// as a serializable object.
+        /// </summary>
+        /// <remarks>
+        /// <para>
+        /// Encrypts a zero plaintext with the public key and returns the ciphertext
+        /// as a serializable object.
+        /// </para>
+        /// <para>
+        /// The encryption parameters for the resulting ciphertext correspond to
+        /// the given ParmsId. Dynamic memory allocations in the process are allocated
+        /// from the memory pool pointed to by the given MemoryPoolHandle.
+        /// </para>
+        /// </remarks>
+        /// <param name="parmsId">The ParmsId for the resulting ciphertext</param>
+        /// <param name="pool">The MemoryPoolHandle pointing to a valid memory
+        /// pool</param>
+        /// <exception cref="ArgumentNullException">if parmsId is null</exception>
+        /// <exception cref="InvalidOperationException">if a public key is not
+        /// set</exception>
+        /// <exception cref="ArgumentException">if parmsId is not valid for the
+        /// encryption parameters</exception>
+        /// <exception cref="ArgumentException">if pool is uninitialized</exception>
+        public Serializable<Ciphertext> EncryptZero(
+            ParmsId parmsId,
+            MemoryPoolHandle pool = null)
+        {
+            Ciphertext destination = new Ciphertext();
+            EncryptZero(parmsId, destination, pool);
+            return new Serializable<Ciphertext>(destination);
+        }
+
+        /// <summary>
+        /// Encrypts a zero plaintext with the public key and returns the ciphertext
+        /// as a serializable object.
+        /// </summary>
+        /// <remarks>
+        /// <para>
+        /// Encrypts a zero plaintext with the public key and returns the ciphertext
+        /// as a serializable object.
+        /// </para>
+        /// <para>
+        /// The encryption parameters for the resulting ciphertext correspond to the
+        /// highest (data) level in the modulus switching chain. Dynamic memory
+        /// allocations in the process are allocated from the memory pool pointed to
+        /// by the given MemoryPoolHandle.
+        /// </para>
+        /// </remarks>
+        /// <param name="pool">The MemoryPoolHandle pointing to a valid memory
+        /// pool</param>
+        /// <exception cref="InvalidOperationException">if a public key is not
+        /// set</exception>
+        /// <exception cref="ArgumentException">if pool is uninitialized</exception>
+        public Serializable<Ciphertext> EncryptZero(MemoryPoolHandle pool = null)
+        {
+            Ciphertext destination = new Ciphertext();
+            EncryptZero(destination, pool);
+            return new Serializable<Ciphertext>(destination);
+        }
+
+        /// <summary>
         /// Encrypts a plaintext with the secret key and stores the result in destination.
         /// </summary>
         /// <remarks>